群组id由int改成long

+++ msip-bhu-api/msip_bhu_rest/src/main/java/com/bhu/vas/web/device/DeviceGroupController.java
@@ -1,14 +1,9 @@
 package com.bhu.vas.web.device;
 
-import java.util.List;
-
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.bhu.vas.api.vto.DeviceGroupVTO;
-import com.smartwork.msip.cores.orm.support.page.TailPage;
-import com.smartwork.msip.jdo.ResponseSuccess;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -16,11 +11,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.bhu.vas.api.rpc.RpcResponseDTO;
-import com.bhu.vas.api.rpc.devices.dto.DeviceGroupDTO;
 import com.bhu.vas.api.rpc.devices.iservice.IDeviceGroupRpcService;
+import com.bhu.vas.api.vto.DeviceGroupVTO;
 import com.bhu.vas.msip.cores.web.mvc.spring.BaseController;
 import com.bhu.vas.msip.cores.web.mvc.spring.helper.SpringMVCHelper;
+import com.smartwork.msip.cores.orm.support.page.TailPage;
 import com.smartwork.msip.jdo.ResponseError;
+import com.smartwork.msip.jdo.ResponseSuccess;
 
 @Controller
 @RequestMapping("/devices/group")
@@ -42,7 +39,7 @@ public class DeviceGroupController extends BaseController{
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@RequestParam(required = true) Integer uid,
-			@RequestParam(required = false,defaultValue="0") int pid,
+			@RequestParam(required = false,defaultValue="0") long pid,
 			@RequestParam(required = false, defaultValue = "1", value = "pn") int pageNo,
 			@RequestParam(required = false, defaultValue = "20", value = "ps") int pageSize) {
 		//RpcResponseDTO<List<DeviceGroupVTO>> birthTree = deviceGroupRpcService.birthTree(uid, pid);
@@ -76,9 +73,9 @@ public class DeviceGroupController extends BaseController{
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@RequestParam(required = true) Integer uid,
-			@RequestParam(required = false,defaultValue="0") int gid,
+			@RequestParam(required = false,defaultValue="0") long gid,
 			@RequestParam(required = true) String name,
-			@RequestParam(required = false,defaultValue="0") int pid
+			@RequestParam(required = false,defaultValue="0") long pid
 			) {
 		System.out.println("~~~~~~~~~~~~save");
 		System.out.println("~~~~~~~~~~~~save:"+deviceGroupRpcService);
@@ -105,7 +102,7 @@ public class DeviceGroupController extends BaseController{
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@RequestParam(required = true) Integer uid,
-			@RequestParam(required = true) int gid,
+			@RequestParam(required = true) long gid,
 			@RequestParam(required = false, defaultValue = "1", value = "pn") int pageNo,
 			@RequestParam(required = false, defaultValue = "20", value = "ps") int pageSize) {
 		RpcResponseDTO<DeviceGroupVTO> detail = deviceGroupRpcService.detail(uid, gid, pageNo, pageSize);
@@ -148,7 +145,7 @@ public class DeviceGroupController extends BaseController{
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@RequestParam(required = true) Integer uid,
-			@RequestParam(required = true) Integer gid,
+			@RequestParam(required = true) Long gid,
 			@RequestParam(required = true) String wifi_ids,
 			@RequestParam(required = true) String group_ids) {
 		RpcResponseDTO<Boolean> grant = deviceGroupRpcService.grant(uid, gid, wifi_ids, group_ids);
@@ -171,7 +168,7 @@ public class DeviceGroupController extends BaseController{
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@RequestParam(required = true) Integer uid,
-			@RequestParam(required = true) int gid,
+			@RequestParam(required = true) long gid,
 			@RequestParam(required = true) String wifi_ids,
 			@RequestParam(required = true) String group_ids) {
 		RpcResponseDTO<Boolean> ungrant = deviceGroupRpcService.ungrant(uid, gid, wifi_ids, group_ids);
