+++ msip-bhu-business-impl/msip_bhu_dataimport/src/main/java/com/bhu/vas/di/op/BuilderWifiDeviceIndexOp.java
@@ -88,7 +88,7 @@ public class BuilderWifiDeviceIndexOp {
 					}
 					//long count = WifiDeviceHandsetPresentSortedSetService.getInstance().presentNotOfflineSize(wifi_mac);
 					long count = WifiDeviceHandsetPresentSortedSetService.getInstance().presentOnlineSize(wifi_mac);
-					List<Integer> groupids = wifiDeviceGroupRelationService.getDeviceGroupIds(wifi_mac);
+					List<Long> groupids = wifiDeviceGroupRelationService.getDeviceGroupIds(wifi_mac);
 					//System.out.println(wifi_mac+"-"+count);
 					indexDto = IndexDTOBuilder.builderWifiDeviceIndexDTO(device, groupids);
 					indexDto.setOnline(device.isOnline() ? 1 : 0);
