+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/dto/search/WifiDeviceIndexDTO.java
@@ -41,7 +41,7 @@ public class WifiDeviceIndexDTO {
 	//wifi设备是否在线
 	private int online;
 	//设备所属群组ids
-	private List<Integer> groupids;
+	private List<Long> groupids;
 	//设备是否为新版本设备
 	private int nvd;
 	//接入的移动设备数量
@@ -161,16 +161,16 @@ public class WifiDeviceIndexDTO {
 	public void setOnline(int online) {
 		this.online = online;
 	}
-	public List<Integer> getGroupids() {
+	public List<Long> getGroupids() {
 		return groupids;
 	}
-	public void setGroupids(List<Integer> groupids) {
+	public void setGroupids(List<Long> groupids) {
 		this.groupids = groupids;
 	}
 	public String getGroups(){
 		if(groupids != null && !groupids.isEmpty()){
 			StringBuffer groups = new StringBuffer();
-			for(Integer groupid : groupids){
+			for(Long groupid : groupids){
 				if(groups.length() > 0) groups.append(StringHelper.WHITESPACE_STRING_GAP);
 				groups.append(String.valueOf(groupid));
 			}
